+++ src/framework/theme/components/calendar/calendar-range.component.ts
@@ -9,6 +9,7 @@ import { Component, EventEmitter, Input, Output, Type } from '@angular/core';
 import { NbCalendarCell, NbCalendarSize, NbCalendarViewMode } from '../calendar-kit/model';
 import { NbDateService } from '../calendar-kit/services/date.service';
 import { NbCalendarRangeDayCellComponent, NbCalendarRangeYearCellComponent } from './calendar-range-cells';
+import { convertToBoolProperty } from '../helpers';
 
 
 export interface NbCalendarRange<D> {
@@ -65,6 +66,7 @@ export interface NbCalendarRange<D> {
  * calendar-header-sub-title-text-line-height:
  * calendar-navigation-button-width:
  * calendar-cell-inactive-text-color:
+ * calendar-cell-in-range-background-color:
  * calendar-cell-disabled-background-color:
  * calendar-cell-disabled-text-color:
  * calendar-cell-selected-background-color:
@@ -93,18 +95,30 @@ export interface NbCalendarRange<D> {
  * calendar-month-cell-height:
  * calendar-year-cell-width:
  * calendar-year-cell-height:
- * calendar-weekday-width:
- * calendar-weekday-height:
+ * calendar-weekday-background:
+ * calendar-weekday-divider-color:
  * calendar-weekday-text-color:
  * calendar-weekday-text-font-size:
  * calendar-weekday-text-font-weight:
  * calendar-weekday-text-line-height:
  * calendar-weekday-holiday-text-color:
- * calendar-cell-in-range-background-color:
+ * calendar-weekday-height:
+ * calendar-weekday-width:
+ * calendar-weeknumber-background:
+ * calendar-weeknumber-divider-color:
+ * calendar-weeknumber-divider-width:
+ * calendar-weeknumber-text-color:
+ * calendar-weeknumber-text-font-size:
+ * calendar-weeknumber-text-font-weight:
+ * calendar-weeknumber-text-line-height:
+ * calendar-weeknumber-height:
+ * calendar-weeknumber-width:
  * calendar-large-width:
  * calendar-large-body-height:
  * calendar-day-cell-large-width:
  * calendar-day-cell-large-height:
+ * calendar-weekday-large-height:
+ * calendar-weekday-large-width:
  * calendar-month-cell-large-width:
  * calendar-month-cell-large-height:
  * calendar-year-cell-large-width:
@@ -127,6 +141,8 @@ export interface NbCalendarRange<D> {
       [visibleDate]="visibleDate"
       [showHeader]="showHeader"
       [size]="size"
+      [showWeekNumber]="showWeekNumber"
+      [weekNumberSymbol]="weekNumberSymbol"
     ></nb-base-calendar>
   `,
 })
@@ -201,6 +217,24 @@ export class NbCalendarRangeComponent<D> {
    * Range which will be rendered as selected.
    * */
   @Input() range: NbCalendarRange<D>;
+
+  /**
+   * Determines should we show week numbers column.
+   * False by default.
+   * */
+  @Input()
+  get showWeekNumber(): boolean {
+    return this._showWeekNumber;
+  }
+  set showWeekNumber(value: boolean) {
+    this._showWeekNumber = convertToBoolProperty(value);
+  }
+  protected _showWeekNumber: boolean = false;
+
+  /**
+   * Sets symbol used as a header for week numbers column
+   * */
+  @Input() weekNumberSymbol: string = '#';
 
   /**
    * Emits range when start selected and emits again when end selected.